+++ src_3_0_0/modules/wfsSearch/tests/unit/components/WfsSearch.spec.js
@@ -12,7 +12,7 @@ const localVue = createLocalVue();
 localVue.use(Vuex);
 config.mocks.$t = key => key;
 
-describe("src/modules/tools/wfsSearch/components/WfsSearch.vue", () => {
+describe("src/modules/wfsSearch/components/WfsSearch.vue", () => {
     const arbitraryFeature = {
         getGeometryName: () => "Klein bottle"
     };
@@ -38,9 +38,10 @@ describe("src/modules/tools/wfsSearch/components/WfsSearch.vue", () => {
         store = new Vuex.Store({
             namespaces: true,
             modules: {
-                Tools: {
+                Modules: {
                     namespaced: true,
                     modules: {
+                        namespaced: true,
                         WfsSearch: WfsSearchModule
                     }
                 },
@@ -55,12 +56,12 @@ describe("src/modules/tools/wfsSearch/components/WfsSearch.vue", () => {
                 uiStyle: sinon.stub()
             }
         });
-        store.commit("Tools/WfsSearch/setActive", true);
+        store.commit("Modules/WfsSearch/setActive", true);
     });
     afterEach(sinon.restore);
 
     it("renders a literal", () => {
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setInstances", instances);
         const wrapper = mount(WfsSearch, {
             localVue,
             store
@@ -70,7 +71,7 @@ describe("src/modules/tools/wfsSearch/components/WfsSearch.vue", () => {
     });
     it("renders multiple literals if configured", () => {
         instances[0].literals.push({}, {});
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setInstances", instances);
         const wrapper = mount(WfsSearch, {
             localVue,
             store
@@ -83,100 +84,100 @@ describe("src/modules/tools/wfsSearch/components/WfsSearch.vue", () => {
             title: "Test WfsSearch II",
             literals: [{}]
         });
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setInstances", instances);
         const wrapper = mount(WfsSearch, {
             localVue,
             store
         });
 
-        expect(wrapper.find("#tool-wfsSearch-instances-select-label").exists()).to.be.true;
-        expect(wrapper.find("#tool-wfsSearch-instances-select").exists()).to.be.true;
+        expect(wrapper.find("#module-wfsSearch-instances-select-label").exists()).to.be.true;
+        expect(wrapper.find("#module-wfsSearch-instances-select").exists()).to.be.true;
     });
     it("renders a container with userHelp if configured", () => {
-        store.commit("Tools/WfsSearch/setUserHelp", "test");
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setUserHelp", "test");
+        store.commit("Modules/WfsSearch/setInstances", instances);
         const wrapper = mount(WfsSearch, {
             localVue,
             store
         });
 
-        expect(wrapper.find("#tool-wfsSearch-userHelp").exists()).to.be.true;
-        expect(wrapper.find("#tool-wfsSearch-userHelp-icon").exists()).to.be.true;
-        expect(wrapper.find("#tool-wfsSearch-userHelp-text").exists()).to.be.true;
+        expect(wrapper.find("#module-wfsSearch-userHelp").exists()).to.be.true;
+        expect(wrapper.find("#module-wfsSearch-userHelp-icon").exists()).to.be.true;
+        expect(wrapper.find("#module-wfsSearch-userHelp-text").exists()).to.be.true;
     });
     it("renders a button to reset the UI", () => {
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setInstances", instances);
         const wrapper = mount(WfsSearch, {
                 localVue,
                 store
             }),
-            resetButton = wrapper.find("#tool-wfsSearch-button-resetUI");
+            resetButton = wrapper.find("#module-wfsSearch-button-resetUI");
 
         expect(resetButton.exists()).to.be.true;
         expect(resetButton.text()).to.equal("common:modules.tools.wfsSearch.resetButton");
     });
     it("renders an input element of type submit to search", () => {
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setInstances", instances);
         const wrapper = mount(WfsSearch, {
                 localVue,
                 store
             }),
-            searchInput = wrapper.find("#tool-wfsSearch-button-search");
+            searchInput = wrapper.find("#module-wfsSearch-button-search");
 
         expect(searchInput.exists()).to.be.true;
         expect(searchInput.element.value).to.equal("common:modules.tools.wfsSearch.searchButton");
         expect(searchInput.element.type).to.equal("submit");
     });
     it("renders a clickable button to show the search results if the user searched and results were found", () => {
-        store.commit("Tools/WfsSearch/setSearched", true);
-        store.commit("Tools/WfsSearch/setResults", [{}]);
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setSearched", true);
+        store.commit("Modules/WfsSearch/setResults", [{}]);
+        store.commit("Modules/WfsSearch/setInstances", instances);
         const wrapper = mount(WfsSearch, {
                 localVue,
                 store
             }),
-            searchButton = wrapper.find("#tool-wfsSearch-button-showResults");
+            searchButton = wrapper.find("#module-wfsSearch-button-showResults");
 
         expect(searchButton.exists()).to.be.true;
         expect(searchButton.text()).to.equal("common:modules.tools.wfsSearch.showResults (1)");
         expect(searchButton.element.disabled).to.be.false;
     });
     it("renders a disabled button if the user searched and no results were found", () => {
-        store.commit("Tools/WfsSearch/setSearched", true);
-        store.commit("Tools/WfsSearch/setResults", []);
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setSearched", true);
+        store.commit("Modules/WfsSearch/setResults", []);
+        store.commit("Modules/WfsSearch/setInstances", instances);
         const wrapper = mount(WfsSearch, {
                 localVue,
                 store
             }),
-            searchButton = wrapper.find("#tool-wfsSearch-button-showResults");
+            searchButton = wrapper.find("#module-wfsSearch-button-showResults");
 
         expect(searchButton.exists()).to.be.true;
         expect(searchButton.text()).to.equal("common:modules.tools.wfsSearch.showResults (0)");
         expect(searchButton.element.disabled).to.be.true;
     });
     it("renders no button if the user searched but the parameter 'resultList' was not configured", () => {
-        store.commit("Tools/WfsSearch/setSearched", true);
-        store.commit("Tools/WfsSearch/setResults", [{}]);
+        store.commit("Modules/WfsSearch/setSearched", true);
+        store.commit("Modules/WfsSearch/setResults", [{}]);
         delete instances[0].resultList;
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setInstances", instances);
         const wrapper = mount(WfsSearch, {
                 localVue,
                 store
             }),
-            searchButton = wrapper.find("#tool-wfsSearch-button-showResults");
+            searchButton = wrapper.find("#module-wfsSearch-button-showResults");
 
         expect(searchButton.exists()).to.be.false;
     });
     it("renders a pagination when more results than are to be shown are available", () => {
-        store.commit("Tools/WfsSearch/setSearched", true);
-        store.commit("Tools/WfsSearch/setResultsPerPage", 2);
-        store.commit("Tools/WfsSearch/setShowResultList", true);
-        store.commit("Tools/WfsSearch/setResults", [
+        store.commit("Modules/WfsSearch/setSearched", true);
+        store.commit("Modules/WfsSearch/setResultsPerPage", 2);
+        store.commit("Modules/WfsSearch/setShowResultList", true);
+        store.commit("Modules/WfsSearch/setResults", [
             arbitraryFeature, arbitraryFeature, arbitraryFeature,
             arbitraryFeature, arbitraryFeature
         ]);
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setInstances", instances);
 
         const wrapper = mount(WfsSearch, {
                 localVue,
@@ -188,14 +189,14 @@ describe("src/modules/tools/wfsSearch/components/WfsSearch.vue", () => {
         expect(pagination.findAll("li").length).to.equal(3);
     });
     it("doesn't render a pagination when 0 is chosen for 'resultsPerPage'", () => {
-        store.commit("Tools/WfsSearch/setSearched", true);
-        store.commit("Tools/WfsSearch/setResultsPerPage", 0);
-        store.commit("Tools/WfsSearch/setShowResultList", true);
-        store.commit("Tools/WfsSearch/setResults", [
+        store.commit("Modules/WfsSearch/setSearched", true);
+        store.commit("Modules/WfsSearch/setResultsPerPage", 0);
+        store.commit("Modules/WfsSearch/setShowResultList", true);
+        store.commit("Modules/WfsSearch/setResults", [
             arbitraryFeature, arbitraryFeature, arbitraryFeature,
             arbitraryFeature, arbitraryFeature
         ]);
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setInstances", instances);
 
         const wrapper = mount(WfsSearch, {
                 localVue,
@@ -206,14 +207,14 @@ describe("src/modules/tools/wfsSearch/components/WfsSearch.vue", () => {
         expect(pagination.exists()).to.be.false;
     });
     it("doesn't render a pagination when resultsPerPage is larger than result list length", () => {
-        store.commit("Tools/WfsSearch/setSearched", true);
-        store.commit("Tools/WfsSearch/setResultsPerPage", 9001);
-        store.commit("Tools/WfsSearch/setShowResultList", true);
-        store.commit("Tools/WfsSearch/setResults", [
+        store.commit("Modules/WfsSearch/setSearched", true);
+        store.commit("Modules/WfsSearch/setResultsPerPage", 9001);
+        store.commit("Modules/WfsSearch/setShowResultList", true);
+        store.commit("Modules/WfsSearch/setResults", [
             arbitraryFeature, arbitraryFeature, arbitraryFeature,
             arbitraryFeature, arbitraryFeature
         ]);
-        store.commit("Tools/WfsSearch/setInstances", instances);
+        store.commit("Modules/WfsSearch/setInstances", instances);
 
         const wrapper = mount(WfsSearch, {
                 localVue,