+++ src_3_0_0/modules/wfsSearch/tests/unit/components/WfsSearchField.spec.js
@@ -10,16 +10,17 @@ const localVue = createLocalVue();
 localVue.use(Vuex);
 config.mocks.$t = key => key;
 
-describe("src/modules/tools/wfsSearch/components/WfsSearchField.vue", () => {
+describe("src/modules/wfsSearch/components/WfsSearchField.vue", () => {
     let store;
 
     beforeEach(() => {
         store = new Vuex.Store({
             namespaces: true,
             modules: {
-                Tools: {
+                Modules: {
                     namespaced: true,
                     modules: {
+                        namespaced: true,
                         WfsSearch: WfsSearchModule
                     }
                 }