+++ CHANGELOG.md
@@ -378,7 +378,7 @@
         - @vue/test-utils: 1.3.0 to 2.2.3
         - mocha: 10.0.0 to 10.1.0
         - vue-loader: 15.9.8 to 17.0.0
-
+- in wfsSearch the 'field'-property type was renamed to queryType
 ### Removed
 - The following NPM packages are removed:
     - dependencies